+++ packages/platform-server/src/transfer_state.ts
@@ -21,15 +21,20 @@ export const TRANSFER_STATE_SERIALIZATION_PROVIDERS: Provider[] = [{
 
 function serializeTransferStateFactory(doc: Document, appId: string, transferStore: TransferState) {
   return () => {
+    // The `.toJSON` here causes the `onSerialize` callbacks to be called.
+    // These callbacks can be used to provide the value for a given key.
+    const content = transferStore.toJson();
+
     if (transferStore.isEmpty) {
       // The state is empty, nothing to transfer,
       // avoid creating an extra `<script>` tag in this case.
       return;
     }
+
     const script = doc.createElement('script');
     script.id = appId + '-state';
     script.setAttribute('type', 'application/json');
-    script.textContent = escapeHtml(transferStore.toJson());
+    script.textContent = escapeHtml(content);
     doc.body.appendChild(script);
   };
 }
